+++ src/lib.rs
@@ -1,5 +1,7 @@
-mod cli_commands;
-mod cli_prompts;
+mod commands;
 mod process;
+mod utils;
 
-pub use cli_commands::*;
+pub use commands::*;
+pub use process::*;
+pub use utils::*;