+++ CHANGELOG.md
@@ -1,5 +1,13 @@
 # CHANGELOG
 
-## 0.0.0
+## 0.1.2
+
+- chore: optimize the clone log
+
+## 0.1.1
+
+- fix: `.ssh/config` file maybe not exists
+
+## 0.1.0
 
 - Initial Release